feat: enforce minimum length of 3 characters for firstName and lastName fields in user schema
