+++ apps/server/database/dbSchema.sql
@@ -15,8 +15,8 @@ CREATE TABLE "users" (
     "email" VARCHAR(255) UNIQUE NOT NULL,
     "username" VARCHAR(255) UNIQUE NOT NULL,
     "password" VARCHAR(255) NOT NULL,
-    "firstName" VARCHAR(255) CHECK (char_length("firstName") >= 1) NOT NULL,
-    "lastName" VARCHAR(255) CHECK (char_length("lastName") >= 1) NOT NULL,
+    "firstName" VARCHAR(255) CHECK (char_length("firstName") >= 3) NOT NULL,
+    "lastName" VARCHAR(255) CHECK (char_length("lastName") >= 3) NOT NULL,
     "birthDate" DATE NOT NULL,
     "role" "roleEnum" NOT NULL DEFAULT 'user',
     "sex" "sexEnum" NOT NULL DEFAULT 'other',
