+++ apps/server/src/models/user.ts
@@ -9,8 +9,8 @@ export const userSchema = z.object({
   email: z.string().email(),
   username: z.string().min(1).max(30),
   password: z.string().regex(passwordRegex),
-  firstName: z.string().min(1).max(255),
-  lastName: z.string().min(1).max(255),
+  firstName: z.string().min(3).max(255),
+  lastName: z.string().min(3).max(255),
   role: z.nativeEnum($Enums.roleEnum),
   sex: z.nativeEnum($Enums.sexEnum),
   birthDate: z.preprocess((val) => dayjs(String(val)).toDate(), z.date()),
